New test: installation on remote multipathed disk
During initial installation the SUT imports a multipathed iSCSI target disk
from a parallel supportserver job and proceeds to install onto it.

Once the initial installation is complete and the client is ready to reboot
the supportserver adds the new kernel and initrd to its PXE server
configuration (custom_pxeboot.pm).  After this the client is supposed
to boot this kernel and initrd via PXE.  The initrd would then re-import
the remote multipathed system disk and proceed from there.

__FIXME__
   As of now this works only on SLES-15 SP1.  Only there the partitioner
   prefers (for some unknown reason) the imported disk over the local one
   for the system disk role.

   For other products either the partitioner would have to be modified
   accordingly or the default local disk of the SUT would have to be taken
   away in the first place (NUMDISKS set to 0. This is currently NOT
   supported, though).

+++ tests/installation/reboot_after_installation.pm
@@ -17,8 +17,10 @@
 use base 'y2_installbase';
 use strict;
 use warnings;
+use lockapi;
 use testapi;
 use utils;
+use mmapi;
 use power_action_utils 'power_action';
 use Utils::Backends 'has_ttys';
 
@@ -30,7 +32,18 @@ sub run {
         my $svirt = console('svirt');
         $svirt->change_domain_element(os => boot => {dev => 'hd'});
     }
+    if (get_var('SUPPORT_SERVER_PXE_CUSTOMKERNEL')) {
+        # we installed on a remote disk provided by a supportserver job and want
+        # to reboot from this disk after downloading the bootconfig via PXE.
+
+        my $jobid_server = (get_parents())->[0] or die "Unexpectedly no parent job found";
+        # "supportserver: please prepare a PXE menu entry with my new bootconfig"
+        mutex_create("custom_pxe_client_ready", $jobid_server);
+        # Wait until the server responds "Got it" (custom_pxeboot.pm)
+        mutex_wait("custom_pxe_ready", $jobid_server);
+    }
     # Reboot
+    # alt-o is the "OK" button in popup "the system will reboot in X seconds..."
     if (has_ttys) {
         my $count = 0;
         while (!wait_screen_change(sub { send_key 'alt-o' }, undef, similarity_level => 20)) {
@@ -44,7 +57,57 @@ sub run {
         # socket end
         send_key 'alt-o';
     }
-    power_action('reboot', observe => 1, keepconsole => 1, first_reboot => 1);
+    if (get_var('SUPPORT_SERVER_PXE_CUSTOMKERNEL')) {
+        # "Press ESC for boot menu"
+        # Expected: match in about 5 seconds
+        assert_screen("initboot_ESC_prompt", 10);
+        send_key "esc";
+
+        # Expected: the BIOS boot menu featuring a netboot entry
+        # This boot menu appears very quickly after ESC: within 1 s:
+        #
+        # Select boot device:
+        #
+        # 1. DVD/CD ...
+        # 2. Floppy ...
+        # 3. Virtio disk ...
+        # 4. iPXE ...
+        # 5. Legacy option rom
+        #
+        # WARNING: the PXE entry number is _volatile_!
+
+        assert_screen("bios_bootmenu", 10);
+        # Since we don't know which entry is the iPXE entry, each case needs
+        # a dedicated needle with distinct secondary tag
+        # Most probable results: 4, then 5
+        # FIXME: as of now, only needle for 4 exists.
+        if (match_has_tag("bios_bootmenu_pxe_is_4")) {
+            send_key "4";
+        }
+        elsif (match_has_tag("bios_bootmenu_pxe_is_5")) {
+            send_key "5";
+        }
+        elsif (match_has_tag("bios_bootmenu_pxe_is_3")) {
+            send_key "3";
+        }
+        elsif (match_has_tag("bios_bootmenu_pxe_is_2")) {
+            send_key "2";
+        }
+        elsif (match_has_tag("bios_bootmenu_pxe_is_1")) {
+            send_key "1";
+        }
+        else {
+            die "BIOS boot menu: unable to detect PXE entry. Giving up.";
+        }
+        #
+        # Expected now: the PXE boot menu with a "custom kernel" menu entry.
+        # Continuation: in boot_from_pxe.pm
+        # (see lib/main_common.pm, sub load_reboot_tests())
+        # beginning with: assert_screen("pxe-custom-kernel", 40);
+    }
+    else {
+        power_action('reboot', observe => 1, keepconsole => 1, first_reboot => 1);
+    }
 }
 
 1;
